Fixed bug #35570 (segfault when re-using soap client object)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_1@202704 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/soap/php_http.c
@@ -32,7 +32,7 @@ static int get_http_headers(php_stream *socketd,char **response, int *out_size T
 
 static int stream_alive(php_stream *stream  TSRMLS_DC)
 {
-	int socket;
+	long socket;
 	char buf;
 
 	/* maybe better to use:
